==> microCT_Calvarial_defects/00_tif2nrrd_previewQuality.ipynb <==
<span STYLE="font-size:150%"> 
    Convert .tif files to .nrrd
</span>

Docker image: gnasello/slicer-env:2023-10-10 \
Latest update: 10 October 2023

- load image stack in Slicer
- segment mineralized tissue
- compute segmented statistics (volumes)

# Load libraries

import pyslicer as ps
import slicer
from pathlib import Path
import pandas as pd
import glob

# Image input

Write the path of the image file directory and the name of the output files

# this cell is tagged 'parameters'
image_directory_name = 'tif stack'
output_dir_path = 'microCT_volume_preview'

Get first image .tif file

image_file = glob.glob(str(Path(image_directory_name) / '*_0000.tif'))[0]
image_file

path = Path(image_file)

# Remove image numbering _0000, _0001 ...
filename_output = path.stem[:-4]

Make ```segmented_volumes``` folder

output_directory = Path(output_dir_path)

output_directory.mkdir(parents=True, exist_ok=True)

# Load image

## Load stack of image files as a 3D volume into 3D Slicer

Write the [X, Y, Z] voxel spaxing if you want it to be different from metadata.

If ```manual_spacing = False```, it will use default spacing value embedded in the image files

manual_spacing = False
# manual_spacing = [0.0107574, 0.0107574, 0.0107574] # mm

Set the quality resolution of the output volume. 

The default quality is ```'preview'``` and you get it with ```quality = Flase```

quality = 'preview'
# quality = 'half'
# quality = 'full'
#quality = False

masterVolumeNode = ps.load.imagestacks(first_image_file = image_file, 
                                       quality=quality,
                                       spacing=manual_spacing)

Print spacing

## mm
masterVolumeNode.GetSpacing()

Export volume Node as `.nrrd` file

filename_output = 'microCT_volume_preview.nrrd'

slicer.util.exportNode(masterVolumeNode, output_directory / filename_output)
