==> microCT_Calvarial_defects/05_segment_polygonDefects.ipynb <==
<span STYLE="font-size:150%"> 
    Segment polygonal regions from microCT scans
</span>

Docker image: gnasello/slicer-env:2023-10-10 \
Latest update: 11 October 2023

- load image stack in Slicer
- segment mineralized tissue
- compute segmented statistics (volumes)

# Load libraries

import pyslicer as ps
import slicer
from pathlib import Path
import pandas as pd

# Volume input

## Load `.nrrd` file into 3D Slicer

Write the path of the `.nrrd` file and load it to Slicer

# this cell is tagged 'parameters'
volume_file = 'microCT_volume/microCT_volume.nrrd'
output_polygon_dir_path = 'segmented_volumes/polygonDefect'
file_nuclei_segmentation = 'segmented_volumes/Bone.seg.nrrd'
directory_notebook = Path().parent.absolute()
sample_name = directory_notebook.stem
camera_view_file = 'segmented_volumes/camera_view.csv'

path = Path(volume_file)

# Remove image numbering _0000, _0001 ...
filename_output = path.stem[:-4]

masterVolumeNode = slicer.util.loadNodeFromFile(volume_file)

Print spacing

## mm
masterVolumeNode.GetSpacing()

Make ```segmented_volumes``` folder

output_polygonDefect_path = Path(output_polygon_dir_path)

output_polygonDefect_path.mkdir(parents=True, exist_ok=True)

Get norebook directory name

sample_name

## Adjust window/level (brightness/contrast)

Adjust the image window/level from the GUI. It is enough to select a region where the signal is present ([youtube](https://slicer.readthedocs.io/en/latest/user_guide/user_interface.html#adjusting-image-window-level)). 

After, get the Window and Level properties from the displayNode associate to the VolumeNode and apply them programmatically anytime you run again the same image.

For more information on the relationship between window/level and brightness/contrast, see [Window and Level Contrast Enhancement](http://fisica.ciens.ucv.ve/curs/dipcourse/html/one-oper/window-level/front-page.html)

displayNode = masterVolumeNode.GetDisplayNode()
print('displayNode.SetWindow(' + str(displayNode.GetWindow()) + ')')
print('displayNode.SetLevel(' + str(displayNode.GetLevel()) + ')')

# displayNode = masterVolumeNode.GetDisplayNode()
# displayNode.AutoWindowLevelOff()
# displayNode.SetWindow(12415.0)
# displayNode.SetLevel(11271.5)

# Create segmentationNode

## Create segmentation-related nodes

Create segmentation node

segmentationNode = ps.segmentation.segmentationNode(name='Segmentation')

Create temporary segment editor to get access to effects

segmentEditorWidget, segmentEditorNode = ps.segmentation.segmentEditorWidget(segmentationNode = segmentationNode, 
                                                                             masterVolumeNode = masterVolumeNode)

## Load bone segmentation to segmentation node

Load segmentation from .nrrd file as labelmap node slicer util module to load from File [github](https://github.com/Slicer/Slicer/blob/master/Base/Python/slicer/util.py#L341-L344)

# Load segmentation from .seg.nrrd file (includes segment names and colors)

bone_labelmap = slicer.util.loadNodeFromFile(file_nuclei_segmentation, 'VolumeFile', properties={'labelmap':True})
# segmentationNode = slicer.util.loadSegmentation(file_nuclei_segmentation)

Import bone labelmapNode to segmentationNode

slicer.modules.segmentations.logic().ImportLabelmapToSegmentationNode(bone_labelmap, segmentationNode)

Rename segment of the segmentation node 

segmentation = segmentationNode.GetSegmentation()
segment = segmentation.GetNthSegment(0)
segment.SetName("Bone")

Delete bone labelmapNode

slicer.mrmlScene.RemoveNode(bone_labelmap)
# masterVolumeNode.SetDisplayVisibility(1)

## Load defect cylinders in segment node

Load `.stl` files as models to Slicer

cylinder_files = [output_polygonDefect_path / 'innerDefectPolygon.vtk',
                  output_polygonDefect_path / 'defectPolygon.vtk'
                 ]

color_list = [
         (230/255, 75/255, 53/255),
         (77/255, 187/255, 213/255),
        ]

cylinderModel_nodes = []

for i, file in enumerate(cylinder_files):
    
    color = color_list[i]
    
    model = slicer.util.loadModel(file)
    model.SetName('Ingrowth_' + file.stem)
    cylinderModel_nodes.append(model)

    slicer.modules.segmentations.logic().ImportModelToSegmentationNode(model, segmentationNode)

    modelDisplayNode = model.GetDisplayNode()
    modelDisplayNode.SetColor(color[0], color[1], color[2])
    modelDisplayNode.SetOpacity(0.4)

Set segment colors

segments_color = {'Bone':(0.9450980392156862, 0.8392156862745098, 0.5686274509803921), # "Bone" color in Slicer
                  'Ingrowth_innerDefectPolygon':(230/255, 75/255, 53/255),
                  'Ingrowth_defectPolygon':(77/255, 187/255, 213/255)
                  }

ps.segmentation.set_segments_color(segments_color, segmentationNode)

# Operation on segments

## Manual fix of the segmentation

Sometimes it might be necessary to remove speckles at the image boundaries. If so, use the `scissor` tool in the `Segment Editor` before proceeding with the rest of the script. 

## Intersect cylinder segments with bone segment

ps.segmentation.logical_intersect('Ingrowth_innerDefectPolygon', 'Bone', segmentationNode, segmentEditorNode, segmentEditorWidget)
ps.segmentation.logical_intersect('Ingrowth_defectPolygon', 'Bone', segmentationNode, segmentEditorNode, segmentEditorWidget)

# Compute bone ingrowth volumes

Delete `Bone` segment before computing volumes

segmentName = 'Bone'
segmentId = segmentationNode.GetSegmentation().GetSegmentIdBySegmentName(segmentName)
segmentationNode.GetSegmentation().RemoveSegment('1')

Compute segment statistics

stats = ps.segmentation.segment_statistics(segmentationNode)

Store volume of each segment

segment_names = []
volumes_ingrowth_mm3 = []

# Display volume of each segment
for segmentId in stats["SegmentIDs"]:
    
    volume_mm3 = stats[segmentId,"LabelmapSegmentStatisticsPlugin.volume_mm3"]
    volumes_ingrowth_mm3.append(volume_mm3)
    
    segmentName = segmentationNode.GetSegmentation().GetSegment(segmentId).GetName()
    segment_names.append(segmentName)

data_dict = {'segmentName': segment_names, 'volume_mm3':volumes_ingrowth_mm3}
df = pd.DataFrame(data_dict)
df

Add sample name columns to DataFrame

df['sample'] = [sample_name] * len(df. index)
df

outputfile = output_polygonDefect_path / (sample_name + '_volume_ingrowth.csv')
df.to_csv(outputfile, index=False)

# Export segments

## Export all segments to individual `.stl` files

Get closed surface representation of the segment, from [slicer scripting repository](https://slicer.readthedocs.io/en/latest/developer_guide/script_repository.html#export-nodes-warped-by-transform-sequence)

segmentationNode.CreateClosedSurfaceRepresentation()

slicer.modules.segmentations.logic().ExportSegmentsClosedSurfaceRepresentationToFiles(str(output_polygonDefect_path), segmentationNode)

# Visualize segmentations in 3D view

Get viewNode

view = slicer.app.layoutManager().threeDWidget(0).threeDView()
viewNode = view.mrmlViewNode()

# Switch off cube and labels
viewNode.SetAxisLabelsVisible(False)
viewNode.SetBoxVisible(False)

color = (28/255, 29/255, 36/255)
# Set view background to RGB color of choice
viewNode.SetBackgroundColor(color[0], color[1], color[2])
viewNode.SetBackgroundColor2(color[0], color[1], color[2])

# Set Orthographic rendering, which is required to show the ruler in a 3D view
viewNode.SetRenderMode(viewNode.Orthographic)

# Set thick and white ruler
viewNode.SetRulerType(2) # 2 - thick
viewNode.SetRulerColor(0) # 0 - white

# Get camera position
cameraNode = slicer.modules.cameras.logic().GetViewActiveCameraNode(viewNode)

## Get camera position in 3D view

Load `camera_view.csv` file

df_camera = pd.read_csv(camera_view_file)
df_camera

## Set lighting in 3D view

slicer.util.selectModule('Lights')
moduleWidget = slicer.modules.lights.widgetRepresentation().self()

Set the Managed 3D view

moduleWidget.logic.addManagedView(viewNode)

Set Key light parameters

LightIntensity = 0.92
LightWarmth = 0.36
LightElevation = 47.45
LightAzimuth = 16.50

moduleWidget.logic.lightKit.SetKeyLightIntensity(LightIntensity)
moduleWidget.logic.lightKit.SetKeyLightWarmth(LightWarmth)
moduleWidget.logic.lightKit.SetKeyLightElevation(LightElevation)
moduleWidget.logic.lightKit.SetKeyLightAzimuth(LightAzimuth)

moduleWidget.updateWidgetFromLogic()

## Set camera position in 3D view

Choose pre-set values for camera position, focal point and view up direction

cameraNode.SetPosition(df_camera['position'])
cameraNode.SetViewAngle(df_camera['viewAngle'][0])
cameraNode.SetViewUp(df_camera['viewUp'])
cameraNode.SetFocalPoint(df_camera['focalPoint'])
cameraNode.SetParallelScale(df_camera['parallelScale'][0])
