==> microCT_Calvarial_defects/07_calvaria_3Dvisualizer.ipynb <==
<span STYLE="font-size:150%"> 
    Slicer 3D visualization 
</span>

Docker image: gnasello/slicer-env:2024-05-08.2 \
Latest update: 8 May 2024

- load .stl surface model files obtained from segmentations
- visualize models in 3D

Load libraries

import slicer
import pyslicer as ps
import pandas as pd
from pathlib import Path

def calvarial_3Dvisualizer(model_dir):

    # Switch to "One 3D view" layout
    layoutManager = slicer.app.layoutManager()
    layoutManager.setLayout(slicer.vtkMRMLLayoutNode.SlicerLayoutOneUp3DView)
    
    slicer.mrmlScene.Clear()
    ps.view.default_dark_3D_view()
    
    model_dir_path = Path(model_dir)
    
    model_file = model_dir_path / 'segmented_volumes/Bone.stl'
    camera_file = model_dir_path / 'segmented_volumes/camera_view.csv'
    
    model = ps.model.load(model_file, color=(0.9450980392156862, 0.8392156862745098, 0.5686274509803921)) # "Bone" color in Slicer
    
    df_camera = pd.read_csv(camera_file)
    
    ps.view.set_camera_3Dview(position=df_camera['position'],
                              viewAngle=df_camera['viewAngle'][0],
                              viewUp=df_camera['viewUp'],
                              parallelScale=df_camera['parallelScale'][0])
    
    imagefile = model_dir_path / 'bone_defect.png'
    ps.view.screenshot_3Dview(imagefile)
    
    model_file = model_dir_path / 'segmented_volumes/polygonDefect/defectPolygon.vtk'
    
    second_model = ps.model.load(model_file, color=(77/255, 187/255, 213/255), opacity = 0.4)
    
    imagefile = model_dir_path / 'bone_defect_polygon.png'
    ps.view.screenshot_3Dview(imagefile)
    
    # points_file = model_dir_path / 'segmented_volumes/polygonDefect/F.mrk.json'
    points_file = model_dir_path / 'segmented_volumes/polygonDefect/F.mrk.json'
    
    pointListNode = slicer.util.loadNodeFromFile(points_file, 'MarkupsFile')
    
    imagefile = model_dir_path / 'bone_defectPolygon_points.png'
    ps.view.screenshot_3Dview(imagefile)

    slicer.mrmlScene.RemoveNode(second_model)
    slicer.mrmlScene.RemoveNode(pointListNode)
    
    model_file = model_dir_path / 'segmented_volumes/originalDefect/OriginalDefect.vtk'
    
    second_model = ps.model.load(model_file, color=(216/255, 101/255, 79/255), opacity = 0.4)
    
    imagefile = model_dir_path / 'bone_originalDefect.png'
    ps.view.screenshot_3Dview(imagefile)

Visualize model

# this cell is tagged 'parameters'
model_dir = 'GN048_AS-IL2_1/'

calvarial_3Dvisualizer(model_dir)
